Sadly, the switch-case building code is not easily reusable.
